move `cmp` and `trie` to `internal/`
`cmp` and `trie` are both modules that are depended upon by `routeit`, but are not meant to be shared publicly. Moving them to `internal/` follows the Go convention and means that they cannot be accessed outside of the `routeit` project and its subdirectories
