+++ internal/cmp/exact.go
