+++ internal/cmp/exact_test.go
