+++ internal/cmp/wildcard.go
