+++ internal/cmp/wildcard_test.go
