+++ internal/trie/string_trie_test.go
